Teacher-Attendance Record\Better folder structure.
