+++ frontend/pages/admin/classRelated/ClassDetails.js
